@@ -19,32 +19,37 @@
   },
   "private": true,
   "dependencies": {
+    "@angular/animations": "^9.0.0-rc.6",
     "@angular/common": "^9.0.0-rc.6",
-    "@angular/compiler": "2.4.4",
+    "@angular/compiler": "9.0.0-rc.6",
     "@angular/core": "^9.0.0-rc.6",
-    "@angular/forms": "2.4.4",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "2.4.4",
-    "@angular/platform-browser-dynamic": "2.4.4",
-    "@angular/router": "3.4.5",
+    "@angular/forms": "^9.0.0-rc.6",
+    "@angular/material": "^9.0.0-rc.6",
+    "@angular/platform-browser": "^9.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "^9.0.0-rc.6",
+    "@angular/router": "^9.0.0-rc.6",
     "angular-calendar": "^0.28.2",
     "core-js": "^2.4.1",
+    "date-fns": "^2.8.1",
     "dragula": "^3.7.2",
     "file-saver": "^1.3.3",
-    "jquery": "^2.2.4",
+    "jquery": "^3.4.1",
     "materialize-css": "~0.97.8",
+    "ngx-materialize": "^6.1.3",
     "openvidu-browser": "1.8.0",
     "primeng": "~1.1.4",
     "quill": "~1.1.10",
-    "rxjs": "^5.0.1",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.0.0-rc.0",
     "time-ago-pipe": "~1.1.1",
     "ts-helpers": "^1.1.1",
     "tslib": "^1.10.0",
-    "zone.js": "^0.6.23"
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
+    "@angular-devkit/build-angular": "^0.803.21",
     "@angular/cli": "^9.0.0-rc.6",
-    "@angular/compiler-cli": "^2.4.0",
+    "@angular/compiler-cli": "^9.0.0-rc.6",
     "@types/jasmine": "2.5.41",
     "codelyzer": "~0.0.26",
     "jasmine-core": "2.4.1",
@@ -60,6 +65,6 @@
     "protractor": "4.0.5",
     "ts-node": "1.2.1",
     "tslint": "3.13.0",
-    "typescript": "2.6.1"
+    "typescript": "^3.6.4"
   }
 }
