@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-import { Subscription }             from 'rxjs/Subscription';
+import { Subscription }             from 'rxjs';
 
 import { File }      from '../../classes/file';
 import { FileGroup } from '../../classes/file-group';
