@@ -1,6 +1,6 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthenticationService } from '../../services/authentication.service';
+import {Component, OnInit} from '@angular/core';
+import {Router} from '@angular/router';
+import {AuthenticationService} from '../../services/authentication.service';
 
 @Component({
   selector: 'app-presentation',
@@ -10,13 +10,17 @@ import { AuthenticationService } from '../../services/authentication.service';
 
 export class PresentationComponent implements OnInit {
 
-  constructor(private authenticationService: AuthenticationService, private router: Router) { }
+  constructor(private authenticationService: AuthenticationService, private router: Router) {
+  }
 
   ngOnInit() {
     //If the user is loggedIn, navigates to dashboard
     this.authenticationService.checkCredentials()
-      .then(() => { this.router.navigate(['/courses']); })
-      .catch((e) => { });
+      .then(() => {
+        this.router.navigate(['/courses']);
+      })
+      .catch((e) => {
+      });
   }
 
 }
