@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
-import { Http, RequestOptions, Headers, Response } from '@angular/http';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import 'rxjs/add/operator/map';
 
 import { User } from '../classes/user';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
 
 @Injectable()
 export class AuthenticationService {
@@ -16,23 +16,19 @@ export class AuthenticationService {
   private user: User;
   private role: string;
 
-  constructor(private http: Http, private router: Router) {
+  constructor(private http: HttpClient, private router: Router) {
     //this.reqIsLogged().catch((e) => { });
   }
 
   logIn(user: string, pass: string) {
     let userPass = user + ":" + pass;
-    let headers = new Headers({
+    let headers = new HttpHeaders({
       'Authorization': 'Basic ' + utf8_to_b64(userPass),
       'X-Requested-With': 'XMLHttpRequest'
     });
-    let options = new RequestOptions({ headers });
+    let options = { headers };
 
     return this.http.get(this.urlLogIn, options)
-      .map(response => {
-        this.processLogInResponse(response);
-        return this.user;
-      })
       .catch(error => Observable.throw(error));
   }
 
@@ -60,7 +56,7 @@ export class AuthenticationService {
   }
 
   private processLogInResponse(response) {
-    
+
     // Correctly logged in
     console.log("User is already logged");
 
@@ -87,10 +83,10 @@ export class AuthenticationService {
 
       console.log("Checking if user is logged");
 
-      let headers = new Headers({
+      let headers = new HttpHeaders({
         'X-Requested-With': 'XMLHttpRequest'
       });
-      let options = new RequestOptions({ headers });
+      let options = { headers };
 
       this.http.get(this.urlLogIn, options).subscribe(
         response => { this.processLogInResponse(response); resolve() },
