@@ -1,20 +1,21 @@
-import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, Response, BrowserXhr } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
+import {Injectable} from '@angular/core';
 
 import * as FileSaver from "file-saver";
 
-import { FileGroup }             from '../classes/file-group';
-import { File }                  from '../classes/file';
-import { CourseDetails }         from '../classes/course-details';
-import { AuthenticationService } from './authentication.service';
+import {FileGroup} from '../classes/file-group';
+import {File} from '../classes/file';
+import {CourseDetails} from '../classes/course-details';
+import {AuthenticationService} from './authentication.service';
 
 import 'rxjs/Rx';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+import {Observable} from "rxjs";
 
 @Injectable()
 export class FileService {
 
-  constructor(private http: Http, private authenticationService: AuthenticationService) { }
+  constructor(private http: HttpClient, private authenticationService: AuthenticationService) {
+  }
 
   private url = "/api-files";
 
@@ -26,14 +27,12 @@ export class FileService {
     console.log("POST new filegroup");
 
     let body = JSON.stringify(fileGroup);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.post(this.url + "/" + courseDetailsId, body, options)
-      .map(response => {
-        console.log("POST new filegroup SUCCESS. Response: ", (response.json() as CourseDetails));
-        return (response.json() as CourseDetails);
-      })
-      .catch(error => this.handleError("POST new filegroup FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.post<CourseDetails>(this.url + "/" + courseDetailsId, body, options);
   }
 
   //DELETE existing FileGroup. Requires the fileGroup id and its course's id
@@ -41,14 +40,12 @@ export class FileService {
   public deleteFileGroup(fileGroupId: number, courseId: number) {
     console.log("DELETE filegroup " + fileGroupId);
 
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.delete(this.url + "/delete/file-group/" + fileGroupId + "/course/" + courseId, options)
-      .map(response => {
-        console.log("DELETE filegroup SUCCESS");
-        return (response.json() as FileGroup);
-      })
-      .catch(error => this.handleError("DELETE filegroup FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.delete<FileGroup>(this.url + "/delete/file-group/" + fileGroupId + "/course/" + courseId, options);
   }
 
   //DELETE existing File. Requires the file id, the fileGroup id that owns it and their course's id
@@ -56,14 +53,12 @@ export class FileService {
   public deleteFile(fileId: number, fileGroupId: number, courseId: number) {
     console.log("DELETE file " + fileId);
 
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.delete(this.url + "/delete/file/" + fileId + "/file-group/" + fileGroupId + "/course/" + courseId, options)
-      .map(response => {
-        console.log("DELETE file SUCCESS");
-        return (response.json() as File);
-      })
-      .catch(error => this.handleError("DELETE file FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.delete<File>(this.url + "/delete/file/" + fileId + "/file-group/" + fileGroupId + "/course/" + courseId, options);
   }
 
   //PUT existing FileGroup. Requires the modified FileGroup and the course id
@@ -71,29 +66,26 @@ export class FileService {
   public editFileGroup(fileGroup: FileGroup, courseId: number) {
     console.log("PUT existing filegroup " + fileGroup.id);
     let body = JSON.stringify(fileGroup);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.put(this.url + "/edit/file-group/course/" + courseId, body, options)
-      .map(response => {
-        console.log("PUT existing filegroup SUCCESS. Response: ", (response.json() as FileGroup));
-        return (response.json() as FileGroup);
-      })
-      .catch(error => this.handleError("PUT existing filegroup FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.put<FileGroup>(this.url + "/edit/file-group/course/" + courseId, body, options);
+    //.catch(error => this.handleError("PUT existing filegroup FAIL. Response: ", error));
   }
 
   //PUT 2 FileGroups. Requires the id of the file moved, the ids of the source and the target FileGroups, the id of the Course and the position of the file in the target FileGroup
   //On success returns the all the fileGroups of the course
   public editFileOrder(fileMovedId: number, fileGroupSourceId: number, fileGroupTargetId: number, filePosition: number, courseId: number) {
-    console.log("PUT existing filegroups (editing file order). From " + fileGroupSourceId +  " to " + fileGroupTargetId + " into position " + fileGroupTargetId);
+    console.log("PUT existing filegroups (editing file order). From " + fileGroupSourceId + " to " + fileGroupTargetId + " into position " + fileGroupTargetId);
 
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.put(this.url + "/edit/file-order/course/" + courseId + "/file/" + fileMovedId + "/from/" + fileGroupSourceId + "/to/" + fileGroupTargetId + "/pos/" + filePosition, options)
-      .map(response => {
-        console.log("PUT existing filegroups SUCCESS (edit file order). Response: ", (response.json() as FileGroup[]));
-        return (response.json() as FileGroup[]);
-      })
-      .catch(error => this.handleError("PUT existing filegroups FAIL (edit file order). Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.put<FileGroup[]>(this.url + "/edit/file-order/course/" + courseId + "/file/" + fileMovedId + "/from/" + fileGroupSourceId + "/to/" + fileGroupTargetId + "/pos/" + filePosition, options)
   }
 
   //PUT existing File. Requires the modified File and the course id
@@ -102,14 +94,13 @@ export class FileService {
     console.log("PUT existing file " + file.name);
 
     let body = JSON.stringify(file);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.put(this.url + "/edit/file/file-group/" + fileGroupId + "/course/" + courseId, body, options)
-      .map(response => {
-        console.log("PUT existing file SUCCESS. Response: ", (response.json() as FileGroup));
-        return (response.json() as FileGroup);
-      })
-      .catch(error => this.handleError("PUT existing filegroup FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.put<FileGroup>(this.url + "/edit/file/file-group/" + fileGroupId + "/course/" + courseId, body, options);
+    //.catch(error => this.handleError("PUT existing filegroup FAIL. Response: ", error));
   }
 
   public downloadFile(courseId: number, file: File) {
@@ -130,15 +121,17 @@ export class FileService {
 
     // Xhr callback when we get a result back
     // We are not using arrow function because we need the 'this' context
-    xhr.onreadystatechange = function() {
+    xhr.onreadystatechange = function () {
 
       // We use setTimeout to trigger change detection in Zones
-      setTimeout(() => { self.pendingDownload = false; }, 0);
+      setTimeout(() => {
+        self.pendingDownload = false;
+      }, 0);
 
       // If we get an HTTP status OK (200), save the file using fileSaver
       if (xhr.readyState === 4 && xhr.status === 200) {
         console.log("File download SUCCESS. Response: ", this.response);
-        var blob = new Blob([this.response], { type: this.response.type });
+        var blob = new Blob([this.response], {type: this.response.type});
         FileSaver.saveAs(blob, file.name);
       }
     };
@@ -148,7 +141,7 @@ export class FileService {
   }
 
   private openFile(response) {
-    var blob = new Blob([response._body], { type: 'text/plain' });
+    var blob = new Blob([response._body], {type: 'text/plain'});
     var url = window.URL.createObjectURL(blob);
     window.open(url)
   }
