@@ -1,21 +1,18 @@
-import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, Response } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-
-import { Session } from '../classes/session';
-import { Entry } from '../classes/entry';
-import { Comment } from '../classes/comment';
-import { Forum } from '../classes/forum';
-import { CourseDetails } from '../classes/course-details';
-import { AuthenticationService } from './authentication.service';
+import {Injectable} from '@angular/core';
+import {Entry} from '../classes/entry';
+import {Comment} from '../classes/comment';
+import {AuthenticationService} from './authentication.service';
 
 
 import 'rxjs/Rx';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+import {Observable} from "rxjs";
 
 @Injectable()
 export class ForumService {
 
-  constructor(private http: Http, private authenticationService: AuthenticationService) { }
+  constructor(private http: HttpClient, private authenticationService: AuthenticationService) {
+  }
 
   private urlNewEntry = "/api-entries";
   private urlNewComment = "/api-comments";
@@ -27,14 +24,13 @@ export class ForumService {
     console.log("POST new entry");
 
     let body = JSON.stringify(entry);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.post(this.urlNewEntry + "/forum/" + courseDetailsId, body, options)
-      .map(response => {
-        console.log("POST new entry SUCCESS. Response: ", (response.json()));
-        return (response.json());
-      })
-      .catch(error => this.handleError("POST new entry FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.post<{comment, entry}>(this.urlNewEntry + "/forum/" + courseDetailsId, body, options);
+    //.catch(error => this.handleError("POST new entry FAIL. Response: ", error));
   }
 
   //POST new Comment. Requires a Comment, the id of the Entry that owns it and the id of the CourseDetails that owns the Forum
@@ -43,14 +39,12 @@ export class ForumService {
     console.log("POST new comment");
 
     let body = JSON.stringify(comment);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.post(this.urlNewComment + "/entry/" + entryId + "/forum/" + courseDetailsId, body, options)
-      .map(response => {
-        console.log("POST new comment SUCCESS. Response: ", (response.json()));
-        return (response.json());
-      })
-      .catch(error => this.handleError("POST new comment FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.post<{comment, entry}>(this.urlNewComment + "/entry/" + entryId + "/forum/" + courseDetailsId, body, options);
   }
 
   //PUT existing Forum. Requires a boolean value for activating/deactivating the Forum and the id of the CourseDetails that owns it
@@ -59,14 +53,13 @@ export class ForumService {
     console.log("PUT existing forum " + (activated ? "(activate)" : "(deactivate)"));
 
     let body = JSON.stringify(activated);
-    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer ' + this.authenticationService.token });
-    let options = new RequestOptions({ headers });
-    return this.http.put(this.urlEditForum + "/edit/" + courseDetailsId, body, options)
-      .map(response => {
-        console.log("PUT existing forum SUCCESS. Response: ", (response.json() as boolean));
-        return (response.json() as boolean);
-      })
-      .catch(error => this.handleError("PUT existing forum FAIL. Response: ", error));
+    let headers = new HttpHeaders({
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + this.authenticationService.token
+    });
+    let options = ({headers});
+    return this.http.put<boolean>(this.urlEditForum + "/edit/" + courseDetailsId, body, options);
+    //.catch(error => this.handleError("PUT existing forum FAIL. Response: ", error));
   }
 
   private handleError(message: string, error: any) {
