@@ -1,10 +1,9 @@
 import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
-import { Http, Headers, RequestOptions, Response } from '@angular/http';
-import { Observable } from "rxjs/Observable";
 
 import { Session } from '../classes/session';
 import { Course } from '../classes/course';
+import {HttpClient, HttpHeaders} from "@angular/common/http";
+import {Observable} from "rxjs";
 
 @Injectable()
 export class VideoSessionService {
@@ -14,17 +13,11 @@ export class VideoSessionService {
 
   private urlSessions = '/api-video-sessions';
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   getSessionIdAndToken(mySessionId) {
     console.log("Getting OpenVidu sessionId and token for lesson '" + mySessionId + "'");
-
-    return this.http.get(this.urlSessions + "/get-sessionid-token/" + mySessionId)
-      .map(response => {
-        console.log("OpenVidu sessionId and token retrieval SUCCESS. Response: ", response);
-        return (response.json());
-      })
-      .catch(error => this.handleError("ERROR getting OpenVidu sessionId and token: ", error));
+    return this.http.get(this.urlSessions + "/get-sessionid-token/" + mySessionId);
   }
 
   removeUser(sessionName) {
@@ -33,14 +26,9 @@ export class VideoSessionService {
     let jsonBody = JSON.stringify({
       'lessonId': sessionName
     });
-    let headers = new Headers({ 'Content-Type': 'application/json' });
-    let options = new RequestOptions({ headers });
-    return this.http.post(this.urlSessions + "/remove-user", jsonBody, options)
-      .map(response => {
-        console.log("User removed from session succesfully. Response: ", response.text());
-        return (response.text());
-      })
-      .catch(error => this.handleError("ERROR removing user from session: ", error));
+    let headers = new HttpHeaders({ 'Content-Type': 'application/json' });
+    let options = ({ headers });
+    return this.http.post(this.urlSessions + "/remove-user", jsonBody, options);
   }
 
   private handleError(message: string, error: any) {
